Bug with broken images in html and epub is fixed
This fix in a certain way cancels #116, but it's necessary. The fact is
that during making html and epub output all the images are supposed to
be directly in chapter folders. Commands that link or copy images don't
descent into subdirectories, so images in "assets" folder are not copied
to destination. So, there was two ways to fix this: first, to change
logic to recursively walk into all subfolders and find all images, and
second, to place all images in some uniform way (i.e. only in chapter
directories, along with text files). The latter was much more easier to
make.