Commas added to contributors list
In output files the list was formed as a solid stream of words, without
line breaks. One solution is to add two spaces in the end of each line.
It's the way to introduce line break when converting from md to other
formats. Another solution is to put some punctuation, which is more
visual for editing source md. I've chosen the latter.